Change for-loop to filter call.
Change the for-loop that may cause a performance issue as mentioned
in the previous commit (SHA: 45e874) to a call to filter() on the whole
Array as suggested here: http://stackoverflow.com/a/24382646/4600952
This looks like the "Swiftier" way to do this right now, but whether
it's better performance wise is unclear.

We may want to benchmark the different options if we're really
concerned about performance.